+++ src/actions/adjointbc/mod_chidg_adjointbc.f90
@@ -238,8 +238,8 @@ contains
 
 
         ! Delete tmp files
-        call delete_file(tmp_flow_file)
-        call delete_file(tmp_adj_file)
+        if (IRANK == GLOBAL_MASTER) call delete_file(tmp_flow_file)
+        if (IRANK == GLOBAL_MASTER) call delete_file(tmp_adj_file)
 
 
     end subroutine chidg_adjointbc
